test hash

+++ symtablehash.c
@@ -157,7 +157,7 @@ void SymTable_expand(SymTable_T oSymTable)
 {
     struct Binding **new_buckets;
     size_t i = 0;
-    if (bindex == ((int)sizeof(bucket_cnts)/(int)sizeof(int)) - 1)
+    if (bindex == (sizeof(bucket_cnts)/sizeof(size_t)) - 1)
     {
         return;
     }
